Hit most number of balloons
You are given an infinite two-dimensional grid. There are N balloons placed at certain coordinates of the grid. You have an arrow with yourself, which you will be using to shoot the balloons. You can select any point on the grid as your starting point and any point on the grid as the target point. When you fire the arrow, all balloons lying on the shortest path between the starting point and target point will be burst. Given the coordinates of the N balloons in an array arr, your task is to find out the maximum number of balloons that you can fire in one arrow shot.