Operations in queue
Given a queue of integers and Q queries. The task is to perform operations on queue according to the query. 

Queries are as:

i x : (adds element x in the queue from rear).

r : (Removes the front element of queue).

h : (Returns the front element).

f y : (check if the element y is present or not in the queue). Return "Yes" if present, else "No".